requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12
celery>=5.3
redis>=4.0